requirements.txt
pyyaml==6.0.1
psycopg2-binary==2.9.9
google-api-python-client==2.149.0
youtube-transcript-api==0.6.2
anthropic==0.34.2
tweepy==4.14.0
feedparser==6.0.11
beautifulsoup4==4.12.3
schedule==1.2.2
pytest==8.3.3
python-dotenv==1.0.1
httpx==0.27.2
tenacity==9.0.0
